Migrate to GHA for builds
