--- examples/.gitignore
@@ -1,4 +1,0 @@
-.pulumi/
-**/bin/
-node_modules/
-
