+++ examples/examples_test.go
@@ -2,59 +2,16 @@ package examples
 
 import (
 	"os"
-	"path"
-	"path/filepath"
 	"testing"
 
 	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
 )
-
-
-func TestAccRepoTs(t *testing.T) {
-	test := getJSBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir: path.Join(getCwd(t), "repo", "ts"),
-		})
-
-	integration.ProgramTest(t, &test)
-}
-
-func TestAccRepoPython(t *testing.T) {
-	test := getPythonBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir: path.Join(getCwd(t), "repo", "python"),
-		})
-
-	integration.ProgramTest(t, &test)
-}
 
 func getBaseOptions(t *testing.T) integration.ProgramTestOptions {
 	checkGitHubToken(t)
 	return integration.ProgramTestOptions{
 		ExpectRefreshChanges: true,
 	}
-}
-
-func getJSBaseOptions(t *testing.T) integration.ProgramTestOptions {
-	base := getBaseOptions(t)
-	baseJS := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			"@pulumi/github",
-		},
-	})
-
-	return baseJS
-}
-
-func getPythonBaseOptions(t *testing.T) integration.ProgramTestOptions {
-	base := getBaseOptions(t)
-	basePy := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			filepath.Join("..", "sdk", "python", "bin"),
-		},
-	})
-
-	return basePy
 }
 
 func checkGitHubToken(t *testing.T) {
