--- examples/repo/python/.gitignore
@@ -1,2 +1,0 @@
-*.pyc
-venv/
